pyproject.toml
[project]
name = "katib"
version = "1.0.0-alpha.0"
description = "Bilingual (EN + AR + bilingual) print-grade PDF document generation skill for Claude Code"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Jasem Al Neaimi", email = "jneaimi@gmail.com" }]
requires-python = ">=3.11"
dependencies = [
    "weasyprint>=60.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "platformdirs>=4.0",
    "jsonschema>=4.20",
    "pypdf>=4.0",
    "pillow>=10.0",
]

[project.optional-dependencies]
screenshots = [
    "playwright>=1.44",
]
gemini = [
    "google-genai>=0.5",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "playwright>=1.44",
    "google-genai>=0.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = ["."]
